// src/api/generated/models/suppliersListParams.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Bookora API
 * Contract OpenAPI chính thức cho backend Bookora. Frontend dùng cookie credentials, cookie accessToken/refreshToken và header X-CSRF-Token cho mutation cần CSRF.
 * OpenAPI spec version: 1.0
 */
import type { SuppliersListSortBy } from './suppliersListSortBy';
import type { SuppliersListSortOrder } from './suppliersListSortOrder';
import type { SuppliersListUsageStatus } from './suppliersListUsageStatus';

export type SuppliersListParams = {
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
search?: string;
usageStatus?: SuppliersListUsageStatus;
createdFrom?: string;
createdTo?: string;
sortOrder?: SuppliersListSortOrder;
sortBy?: SuppliersListSortBy;
hasPhone?: boolean;
hasEmail?: boolean;
};